+++ src/main/java/tslib/model/statespace/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Compact state-space and Kalman-filter based forecasting models.
+ */
+package tslib.model.statespace;