+++ src/main/java/tslib/transform/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Data transformation and differencing utilities for time-series preprocessing.
+ */
+package tslib.transform;